requirements.txt
ultralytics
git+https://github.com/facebookresearch/segment-anything-2.git
streamlit
opencv-python-headless
torch
torchvision
sahi
reportlab
pandas
numpy
Pillow
fastapi
uvicorn
httpx
pyyaml
grad-cam
python-multipart
